requirements.txt
flask
flask-cors
beautifulsoup4
requests
python-dotenv
lxml
google-genai
